Move UpdateTest class to session package
