+++ drools-compiler/src/test/resources/org/drools/compiler/integrationtests/session/update_test.drl
@@ -13,7 +13,7 @@
  * limitations under the License.
 */
 
-package org.drools.compiler.integrationtests;
+package org.drools.compiler.integrationtests.session;
 
 import org.drools.core.test.model.Person;
 import org.drools.core.test.model.Cheese;
